backend/requirements.txt
fastapi
uvicorn[standard]
python-jose
passlib[bcrypt]
pydantic
pydantic-settings
python-multipart
python-dateutil
python-dotenv
prisma
pandas
